c261: start api/client.ts split — extract BandMask family + _http (#441 2.4)
First slice of the 1392-line / 96-export client.ts split called out
in #441 P1 item 2.4. Strategy: extract one family per cycle, leaving
re-exports in client.ts for back-compat so consumer imports don't
break.

New modules:
- frontend/src/api/_http.ts (53 LOC) — shared ApiError + request +
  requestBuffer primitives. Lives under '_http' so the per-family
  modules can call it without circular imports through client.ts.
- frontend/src/api/bandmask.ts (165 LOC) — full BandMask family:
  9 types (BandMaskIndex, BandMaskIndexEntry, BandMaskComparisonEntry,
  BandMaskCanonicalComparison, BandMaskHidsagIndex, BandMaskHidsagIndexEntry,
  BandMaskHidsagSummary, BandMaskSummary, HidsagCovariateProbability)
  + a new bandMaskApi runtime object mirroring the existing
  api.bandMasks*() methods. The original api.bandMasks*() entries
  in client.ts remain in place so existing call sites don't churn.

client.ts changes:
- 138 lines of type declarations removed (now in bandmask.ts).
- client.ts: 1392 -> 1234 lines (-158 net after the new import block).
- ApiError + request + requestBuffer imported from _http and
  ApiError re-exported for back-compat.
- BandMask family types re-exported via type-only re-export.

Incidental:
- BandMaskTab.tsx: explicit BandMaskComparisonEntry type annotation
  on a .find() callback. Was previously inferred via the chain; the
  re-export through bandmask.ts triggered tsc-strict to drop the
  inference. Two-character fix.

Verification:
- pnpm typecheck -> clean
- pnpm test -> 30 passed
- pnpm build -> clean

Next slices: api/eda.ts, api/topics.ts, api/hidsag.ts in follow-on
cycles. Eventually api/client.ts becomes a thin barrel + the few
shared types that don't fit a family.
